Module unicorn::messages

Module for stating message contracts and decoding/encoding messages

Example usage:

use unicorn::messages::{common, Msg, encode, decode};

// All messages need an ID. So, create one first.
let id = common::ID {
    uuid: "_test".to_string(),
    component: common::Components::Core,
};

// Create a Heartbeat message
let heartbeat = Msg::Heartbeat { id: id, count: 1 };

// Encode the message to JSON bytes
let jbytes = encode(&heartbeat).unwrap();

// Decode the JSON bytes to original message
let jmsg = decode(jbytes).unwrap();

Modules

common
Common message structures shared by more than one components

Enums

Msg
Message structure definitions. Each variant of this enum forms a type of message.

Functions

decode
Generic decoder for all messages. Decodes JSON strings to message structures

encode
Generic encoder for all messages. Encodes message structures to JSON strings

encode_bytes
Convenience method to message structures directly to serialized bytes