Module unicorn::messages::common

Common message structures shared by more than one components

Structs

ID
Identity message. Used by every message that needs an ID

Enums

Components
List of main component types in unicorn

State
Different states in which a component can be at any given time