Struct unicorn::network::Net

pub struct Net {
    // some fields omitted
}

Network structure that holds network topology information. Aka, the network interface.

Net provides a higher-level abstraction of TCP listener sockets. It multiplexes incoming connections into parallel threads and runs a given Processor against all messages that the connection receives.

Using Net

Net works in 2 steps:

Bind a listener using Net::bind(addr). This starts a TcpListener on the specified address and returns an instance of Net.

Start accepting connections by calling .recv(processor) on the returned Net instance.

Methods

impl Net

fn addr(&self) -> String

Returns the addr

fn status(&self) -> Status

Returns the status

fn num_workers(&mut self, i: usize)

Set number of workers for processing incoming streams in parallel

Note: This method will cause a panic!() if number of workers specified is less than 1.

fn bind(addr: String) -> Result<Net, Error>

Start a listener on an existing network interface

fn recv(&self, processor: &'static Processor)

Process incoming TCP stream